allow all-lines-at-a-time mode
This involves keeping track of whitespace in the tersened_words
iterator so that it can be retained later and printed neatly.

Currently, this doesn't do a whole lot useful, though it does permit
multi-word expansions across lines. I plan to add a re-fold mode, which
is what will actually make it really useful.