refactor main.lua, unbreak frequency CLI
It was spitting out the frequency one line at a time, which was almost
completely useless, due to an ill-advised attempt to reduce code
duplication by calling a function in a spot where all the logic couldn't
be encapsulated in a single function. I inverted that by changing to an
iterator that each of the two code paths can use to retrieve all the
files.

--frequency now implies -o and ignores -w.